requirements.txt
numpy==1.18.5
opencv-python
tqdm
docopt
six
scipy==1.4.1
Pillow
matplotlib
scikit-image==0.16.2
tensorflow==2.2.0
imageio
keras==2.3.1
pandas
drawSvg
imgaug
imagecodecs
notify-run
shapely
psutil

#preprocessor
segments-ai==0.22
imantics
h5py==2.10.0